Refactored to use a basemechtest class
